user stories ok

+++ db/seeds.rb
@@ -1,10 +1,10 @@
 require 'open-uri'
 
 puts "Destroy ingredients"
-Ingredient.destroy_all
+Ingredient.destroy_all if Rails.env.development?
 
 puts "Destroy Cocktails"
-Cocktail.destroy_all
+Cocktail.destroy_all if Rails.env.development?
 
 puts "Create ingredients"
 url = "https://www.thecocktaildb.com/api/json/v1/1/list.php?i=list"
